requirements.txt
# AI Image Generator - GPU Requirements
# For systems with NVIDIA GPU and CUDA support

# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Web Interface
streamlit>=1.28.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
safetensors>=0.4.0

# Progress Tracking
tqdm>=4.66.0

# Additional AI/ML Tools
scipy>=1.11.0
